Move theme classes to client
