+++ WeddingWebsite/Client/Client.csproj
@@ -20,6 +20,8 @@
       <_ContentIncludedByDefault Remove="wwwroot\index.html" />
     </ItemGroup>
 
-
+    <ItemGroup>
+        <Folder Include="Models\" />
+    </ItemGroup>
 
 </Project>
