+++ WeddingWebsite/Client/Models/Theme/BoxStyle.cs
@@ -1,0 +1,6 @@
+﻿namespace WeddingWebsite.Client.Models.Theme;
+
+public sealed record BoxStyle(
+    BoxType Type,
+    SectionTheme InnerTheme
+);
