+++ WeddingWebsite/Client/Models/Theme/SectionTheme.cs
@@ -1,6 +1,4 @@
-﻿using WeddingWebsite.Models.WebsiteConfig;
-
-namespace WeddingWebsite.Models.Theme;
+﻿namespace WeddingWebsite.Client.Models.Theme;
 
 /// <summary>
 /// The theming of a particular section. This can be the same across all sections, or it may vary depending on the section.
