--- WeddingWebsite/Models/Theme/BoxStyle.cs
@@ -1,8 +1,0 @@
-﻿using WeddingWebsite.Models.WebsiteConfig;
-
-namespace WeddingWebsite.Models.Theme;
-
-public sealed record BoxStyle(
-    BoxType Type,
-    SectionTheme InnerTheme
-);
